+++ src/debops/config.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (C) 2020 Maciej Delmanowski <drybjed@gmail.com>
-# Copyright (C) 2020 DebOps <https://debops.org/>
+# Copyright (C) 2020-2023 Maciej Delmanowski <drybjed@gmail.com>
+# Copyright (C) 2020-2023 DebOps <https://debops.org/>
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from .utils import unexpanduser
@@ -204,21 +204,65 @@ class Configuration(object):
         for key, value in os.environ.items():
             print('{}={}'.format(key, value))
 
-    def show(self, format='toml'):
-        if format == 'toml':
-            relative_root = os.path.relpath(os.path.abspath('/'))
-            if self._env_files:
-                print('# Environment files:')
-                for filename in self._env_files:
-                    relative_file = os.path.relpath(unexpanduser(filename))
-                    print('#    ', relative_file.replace(relative_root, '', 1))
-                print()
-            if self._config_files:
-                print('# Configuration files:')
-                for filename in self._config_files:
-                    relative_file = os.path.relpath(unexpanduser(filename))
-                    print('#   ', relative_file.replace(relative_root, '', 1))
-                print()
-            print(toml.dumps(self._config).strip())
-        elif format == 'json':
-            print(json.dumps(self._config, sort_keys=True, indent=4))
+    def show(self):
+        relative_root = os.path.relpath(os.path.abspath('/'))
+        if self._env_files:
+            print('# Environment files:')
+            for filename in self._env_files:
+                relative_file = os.path.relpath(unexpanduser(filename))
+                print(relative_file.replace(relative_root, '', 1))
+            print()
+        if self._config_files:
+            print('# Configuration files:')
+            for filename in self._config_files:
+                relative_file = os.path.relpath(unexpanduser(filename))
+                print(relative_file.replace(relative_root, '', 1))
+
+    def config_get(self, key, format='unix'):
+        key_path = ['.']
+        if key != '.':
+            key_path = list(filter(None, key.split('.')))
+        key_found = True
+
+        _config = self._config
+        if key_path[0] != '.':
+            for element in key_path:
+                try:
+                    _config = _config[element]
+                except KeyError:
+                    key_found = False
+
+        if key_found:
+            if isinstance(_config, dict):
+                if format in ['unix', 'yaml']:
+                    print(yaml.dump(_config).strip())
+                elif format == 'json':
+                    print(json.dumps(_config, sort_keys=True, indent=2))
+                elif format == 'toml':
+                    print(toml.dumps(_config).strip())
+            elif isinstance(_config, bool):
+                if format == 'unix':
+                    print(json.dumps(_config))
+                elif format == 'yaml':
+                    print(yaml.dump(_config).strip())
+                elif format in ['json', 'toml']:
+                    print(json.dumps(_config))
+            elif isinstance(_config, list):
+                if format == 'unix':
+                    for entry in _config:
+                        print(entry)
+                elif format == 'yaml':
+                    print(yaml.dump(_config).strip())
+                elif format in ['json', 'toml']:
+                    print(json.dumps(_config, sort_keys=True))
+            else:
+                if format == 'unix':
+                    print(_config)
+                elif format == 'yaml':
+                    print(yaml.dump(_config).strip())
+                elif format in ['json', 'toml']:
+                    print(json.dumps(_config))
+        else:
+            if format == 'json':
+                print('{}')
+            sys.exit(1)